[recipe] feat:  Add sleep/wakeup mode for gen rm vllm service and add tqdm  showing process (#2739)
### What does this PR do?

Add sleep/wakeup mode for gen rm vllm service and add tqdm showing
process.

This capability is particularly beneficial when the model server shares
resources with a training workload on the same machine. It allows the
reward model service to be temporarily offloaded (to free up GPU memory)
during intensive training sessions and reloaded when the service is
required again.